Mark UIStackView bug as fixed in unit test
This test case was previously testing that a bug in UIStackView
(causing view.isHidden to not work as expected) was still present.
However the test is failing, so it seems this bug must have been
fixed. Curiously, the test (before this commit) succeeded if run
on its own, but failed when run as part of the full test suite.
Marking the bug as fixed for now, but leaving in the
`showsInStackView` workaround to be safe.