Work around SwiftFormat deployment target issue (#50)
PRs are failing due to a build warning about the SwiftFormat pod's
deployment target being set to iOS 7.0, which is too low. Given that we
don't actually care about supporting iOS versions that far back, add a
workaround to the Podfile to set all pods' deployment targets equal to
whatever deployment target we have set for Thumbprint as a whole.